requirements.txt
# Deep Learning Framework
# TensorFlow (CPU-only, no GPU support on Windows)
# tensorflow==2.17.1

# PyTorch with CUDA support (RECOMMENDED for GPU)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
opencv-python==4.8.0.74

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# ML Tools
scikit-learn==1.3.0

# Progress Bar
tqdm==4.65.0

# Model Export
onnx==1.14.0
tf2onnx==1.15.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# API Server
flask==3.0.0
flask-cors==4.0.0
